+++ lib/fromarray.rb
@@ -22,12 +22,10 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-# Ruby Stream API main module.
+# A stream implemented based on an Array.
 # Author:: Mihai Andronache (amihaiemil@gmail.com)
-module StreamApi
-  VERSION = '0.1.0'.freeze
-
-  # If it breaks, go to stderr
-  class Error < StandardError
+class FromArray
+  def initialize(array)
+    @array = array
   end
 end
