+++ lib/stream.rb
@@ -22,21 +22,19 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-module StreamApi
-  # A Stream.
-  # Author:: Mihai Andronache (amihaiemil@gmail.com)
-  class Stream
-    # Build a stream from an array.
-    def self.from_array(array)
-      return FromArray.new(array)
-    end
+require_relative 'fromarray.rb'
 
-    # A stream implemented based on an Array.
-    # Author:: Mihai Andronache (amihaiemil@gmail.com)
-    class FromArray
-      def initialize(array)
-        @array = array
-      end
-    end
-end
+# Ruby Stream API main module.
+# Author:: Mihai Andronache (amihaiemil@gmail.com)
+module Stream
+  VERSION = '0.1.0'.freeze
+
+  # If it breaks, go to stderr
+  class Error < StandardError
+  end
+
+  # Build a stream from an array.
+  def self.from_array(array)
+    return FromArray.new(array)
+  end
 end
