+++ test/stream_test.rb
@@ -24,11 +24,17 @@
 require 'test_helper'
 require 'stream.rb'
 
-module ApiTest
+module StreamTest
   class StreamTest < Minitest::Test
+
+    # The Stream module has a version
+    def test_that_it_has_a_version_number
+      refute_nil Stream::VERSION
+    end
+
+    # The Stream module can be created from an array
     def test_that_from_array_works
-      refute_nil StreamApi::Stream.from_array([1, 2, 3])
-      refute_nil StreamApi::Stream::FromArray.new([1, 2, 3])
+      refute_nil Stream.from_array([1, 2, 3])
     end
   end
 end
